+++ src/syscall/src/common.rs
@@ -1,27 +1,27 @@
-pub const SYS_DEBUG: u32 = 0;
+pub const SYS_DEBUG: usize = 0;
 
 //Linux compatible
-pub const SYS_EXIT: u32 = 1;
-pub const SYS_FORK: u32 = 2;
-pub const SYS_READ: u32 = 3;
-pub const SYS_WRITE: u32 = 4;
-pub const SYS_OPEN: u32 = 5;
-pub const SYS_CLOSE: u32 = 6;
-pub const SYS_EXECVE: u32 = 11;
-pub const SYS_LSEEK: u32 = 19;
-pub const SYS_FSTAT: u32 = 28;
-pub const SYS_DUP: u32 = 41;
-pub const SYS_BRK: u32 = 45;
-pub const SYS_GETTIMEOFDAY: u32 = 78;
-pub const SYS_FSYNC: u32 = 118;
-pub const SYS_YIELD: u32 = 158;
+pub const SYS_EXIT: usize = 1;
+pub const SYS_FORK: usize = 2;
+pub const SYS_READ: usize = 3;
+pub const SYS_WRITE: usize = 4;
+pub const SYS_OPEN: usize = 5;
+pub const SYS_CLOSE: usize = 6;
+pub const SYS_EXECVE: usize = 11;
+pub const SYS_LSEEK: usize = 19;
+pub const SYS_FSTAT: usize = 28;
+pub const SYS_DUP: usize = 41;
+pub const SYS_BRK: usize = 45;
+pub const SYS_GETTIMEOFDAY: usize = 78;
+pub const SYS_FSYNC: usize = 118;
+pub const SYS_YIELD: usize = 158;
 
 //Rust Memory
-pub const SYS_ALLOC: u32 = 1000;
-pub const SYS_REALLOC: u32 = 1001;
-pub const SYS_REALLOC_INPLACE: u32 = 1002;
-pub const SYS_UNALLOC: u32 = 1003;
+pub const SYS_ALLOC: usize = 1000;
+pub const SYS_REALLOC: usize = 1001;
+pub const SYS_REALLOC_INPLACE: usize = 1002;
+pub const SYS_UNALLOC: usize = 1003;
 
 //Misc
-pub const SYS_TIME: u32 = 3000;
-pub const SYS_FPATH: u32 = 3001;
+pub const SYS_TIME: usize = 3000;
+pub const SYS_FPATH: usize = 3001;
